refactor: centralize config dataclasses and transformer
